Update client.py
In the event `res` returns an error, there would be no key with the value `success` in the returned dict.  This results in a KeyError which doesn't provide the caller any information about the reason for the error.  Instead calling `get` will simply default to `None` if no `success` key is present - and raise the appropriate error with the response `msg`.